+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#160] Update Redmine to version 6.0.6
+- [#160] Update ACTIVERECORD_SESSION_STORE_PLUGIN_VERSION to 0.3.0
+- [#160] Added zzz_cloudogu_patches Plugin to fix redmine upgrading issues
 
 ## [v5.1.8-3] - 2025-08-06
 ### Security